+++ src/gui/wallet_test.go
@@ -1413,7 +1413,7 @@ func TestWalletNewSeed(t *testing.T) {
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			gateway := &GatewayerMock{}
-			gateway.On("IsWalletAPIDisabled").Return(false)
+			gateway.On("IsWalletAPIEnabled").Return(true)
 
 			endpoint := "/wallet/newSeed"